requirements.txt
websockets>=11.0
Pillow>=9.0.0
pyautogui>=0.9.50